renamed filereader to error_methods
