+++ .rubocop.yml
@@ -12,7 +12,7 @@ Metrics/MethodLength:
 Metrics/AbcSize:
   Max: 50
 Metrics/ClassLength:
-  Max: 150
+  Max: 220
 Metrics/BlockLength:
   ExcludedMethods: ['describe']
   Max: 30
