+++ bin/main
@@ -1,4 +1,4 @@
 #!/usr/bin/env ruby
 
 require 'colorize'
-puts "Hello".red
+puts 'Hello'.red
